Fix stack memory access
According to the 8051 documentation, all stack operations use indirect memory access (which makes sense, you don't want to be getting your stack and SFRs mixed up).

The way I found this: I am emulating an image that uses the high 128bits of memory for the stack, and I had no idea how it ended up in a routine that was supposed to be disabled, turns out the return address in the stack was corrupted, so on a ret, it jumped to somewhere in the middle of the disabled routine.

Without this change, you can expect memory, register and stack corruption whenever the stack pointer is above 0x7f.